% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_tools.R
\name{add_file_in_yaml}
\alias{add_file_in_yaml}
\title{Add an element in yaml}
\usage{
add_file_in_yaml(yml, name, datetime, extand, description)
}
\arguments{
\item{yml}{yaml path.}

\item{name}{File name.}

\item{datetime}{file datetime.}

\item{extand}{file description.}
}
\description{
Add an element in yaml
}
\examples{
\dontrun{
  yml <- "inst/exemple/files_desc.yaml"
  name <- "toto"
  datetime <- "20200831_142715"
  extand <- "csv"
  description <- "My file desc"
  add_file_in_yaml("inst/exemple/files_desc.yaml", "toto", "20200831_142715", "csv")
}

}
