% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_tools.R
\name{delete_config}
\alias{delete_config}
\title{Delete all YAML files}
\usage{
delete_config(
  base_dir,
  config_file = "files_desc.yaml",
  confirm = TRUE,
  dry_run = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{base_dir}{Folder to check for YAML files}

\item{config_file}{YAML files names}

\item{confirm}{With or without confirmation before deleting}

\item{dry_run}{simulation (no deletion)}

\item{verbose}{Display messages}
}
\value{
delete all yaml files
}
\description{
Delete all YAML files
}
\examples{
\dontrun{
\donttest{
delete_config("~/shinydrive/tests", config_file = "files_desc.yaml")
}
}
}
