% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{management_ui}
\alias{management_ui}
\alias{management_server}
\title{File management shiny module.}
\usage{
management_ui(id)

management_server(
  input,
  output,
  session,
  id,
  save_dir = NULL,
  admin_user = TRUE,
  force_desc = FALSE,
  lan = "EN",
  file_translate = read.csv(system.file("translate/translate.csv", package =
    "shinyfilesmanager"), sep = ";", encoding = "UTF-8")
)
}
\arguments{
\item{id}{\code{character}. shiny id to allow multiple instanciation.}

\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{save_dir}{\code{character}. Directory of the files.}

\item{admin_user}{\code{boolean/reactive} (TRUE). Admin user or not.}

\item{force_desc}{\code{boolean/reactive} (FALSE). Force to add an entry description ?}

\item{lan}{\code{character/reactive} ("EN"). Language to be used in the module (FR and EN available... contributions are welcome :)).}

\item{file_translate}{\code{data.frame/reactive} File for translation.}
}
\description{
File management shiny module.
}
\examples{
\dontrun{

if(require(shiny)){
  ui <- fluidPage(
    management_ui(id = "idm")
  )
  server <- function(input, output, session) {
    callModule(module = management_server,
             id = "idm",
             session = session,
             admin_user = TRUE,
             save_dir =  getwd(),
             lan = "FR")
  }
  shinyApp(ui, server)
}

}

}
