% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moddhiny.R
\name{management_ui}
\alias{management_ui}
\title{Ui for file management module}
\usage{
management_ui(
  id,
  lan = "EN",
  tran = fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))
)
}
\arguments{
\item{id}{\code{character}. Shiny id to allow multiple instanciation.}

\item{lan}{\code{character} ("EN"). Language to be used in the module (FR and EN available... contributions are welcome :)).}

\item{tran}{\code{data.frame} (fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))). File for translation.}
}
\description{
Ui part for the file management module, a shiny module for file management.
}
\examples{
\dontrun{

ui <- fluidPage(
  management_ui(id = "idm",
                lan = "EN")
)
server <- function(input, output, session) {
  callModule(module = managment_server,
             id = "idm",
             session = session,
             admin_user = TRUE,
             save_dir =  getwd(),
             lan = "EN")
}
shinyApp(ui, server)

}
}
