% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_management.R
\name{managment_server}
\alias{managment_server}
\title{server for file management module}
\usage{
managment_server(
  input,
  output,
  session,
  save_dir,
  admin_user = TRUE,
  lan = "EN",
  tran = fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))
)
}
\arguments{
\item{input}{shiny input.}

\item{output}{shiny output.}

\item{session}{shiny session.}

\item{save_dir}{directory for init getwd().}

\item{lan}{language for module avialable are FR and EN. Contribution are welcome :).}

\item{tran}{file for translation}

\item{id}{Module id.}
}
\description{
server part for file management module, it's a shiny module used for file gestion.
}
\examples{
\dontrun{
ui <- fluidPage(
  mangement_ui(id = "idm", lan = "EN")
)
server <- function(input, output, session) {
  callModule(module = managment_server,"idm", lan = "EN", session = session,
             admin_user = TRUE, save_dir = getwd())
}
shinyApp(ui, server)

}

}
