% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_management.R
\name{mangement_ui}
\alias{mangement_ui}
\title{Ui for file management module}
\usage{
mangement_ui(
  id,
  lan = "EN",
  tran = fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))
)
}
\arguments{
\item{id}{Module id.}

\item{lan}{language for module avialable are FR and EN. Contribution are welcome :).}

\item{tran}{file for translation}
}
\description{
Ui part for file management module, it's a shiny module used for file gestion.
}
\examples{
\dontrun{
ui <- fluidPage(
  mangement_ui(id = "idm", lan = "EN")
)
server <- function(input, output, session) {
  callModule(module = managment_server,"idm", lan = "EN", session = session,
             admin_user = TRUE, save_dir =  getwd())
}
shinyApp(ui, server)

}

}
