% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_tools.R
\name{supress_file_in_yaml}
\alias{supress_file_in_yaml}
\title{Supress element in yaml}
\usage{
supress_file_in_yaml(yml, name, datetime, extand)
}
\arguments{
\item{yml}{yaml path.}

\item{name}{File name.}

\item{datetime}{file datetime.}

\item{extand}{file extand.}
}
\description{
Supress element in yaml
}
\examples{
\dontrun{
  yml <- "inst/exemple/files_desc.yaml"
  name <- "toto"
  datetime <- "20200831_142715"
  extand <- "csv"
  supress_file_in_yaml("inst/exemple/files_desc.yaml", "toto", "20200831_142715", "csv")
}

}
