% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_tools.R
\name{delete_old_config}
\alias{delete_old_config}
\title{Delete config files older than a specific number of days}
\usage{
delete_old_config(
  base_dir,
  days = 30,
  config_file = "files_desc.yaml",
  dry_run = FALSE
)
}
\arguments{
\item{base_dir}{base folder}

\item{days}{number of days to check the age of files we want to delete}

\item{config_file}{name of YAML files}

\item{dry_run}{Simulation}
}
\value{
delete old config files from folder
}
\description{
Delete config files older than a specific number of days
}
\examples{
delete_old_config("~/shinydrive/tests", days = 60, config_file = "files_desc.yaml")

}
