% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_tools.R
\name{init_config}
\alias{init_config}
\title{Create YAML configuration files recursively}
\usage{
init_config(base_dir, config_file = "files_desc.yaml", file_patterns = NULL)
}
\arguments{
\item{base_dir}{\code{character} Base directory path}

\item{config_file}{\code{character} YAML configuration file name}

\item{file_patterns}{\code{character} File patterns to match}
}
\value{
Invisible count of processed directories
}
\description{
Create YAML configuration files recursively
}
