% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moddhiny.R
\name{managment_server}
\alias{managment_server}
\title{server for file management module}
\usage{
managment_server(
  input,
  output,
  session,
  id,
  save_dir,
  admin_user = TRUE,
  lan = "EN",
  tran = fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{id}{\code{character}. shiny id to allow multiple instanciation.}

\item{save_dir}{\code{character}. Directory of the files.}

\item{admin_user}{\code{boolean} (TRUE). Admin user or not.}

\item{lan}{\code{character} ("EN"). Language to be used in the module (FR and EN available... contributions are welcome :)).}

\item{tran}{\code{data.frame} (fread(system.file("translate/translate.csv", package = "shinyfilesmanager"))). File for translation.}
}
\description{
server part for file management module, a shiny module for file management.
}
\examples{
\dontrun{

ui <- fluidPage(
  management_ui(id = "idm", lan = "EN")
)
server <- function(input, output, session) {
  callModule(module = managment_server,"idm", lan = "EN", session = session,
             admin_user = TRUE, save_dir = getwd())
}
shinyApp(ui, server)

}
}
